Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -4,6 +4,6 @@ int main()
 {
     char inputCharacter;
     scanf("%c", &inputCharacter);
-    printf("ASCII Value: %d\n", (int)inputCharacter);
+    printf("ASCII Value: %d\n", inputCharacter);
     return 0;
 }
